requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Google Cloud Services
google-cloud-vision==3.5.0
google-cloud-storage==2.13.0
google-cloud-firestore==2.13.1
firebase-admin==6.2.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
Pillow==10.1.0
